+++ test/test_qfm_custom_block.rb
@@ -7,17 +7,33 @@ class TestQfmCustomBlock < Minitest::Test
     text = <<~MD
       :::foo bar
       message
+
+      - list1
+      - list2
+
+      ```ruby
+      puts 'hello'
+      ```
+
+      <div>html block</div>
       :::
     MD
-    @doc = QiitaMarker.render_doc(text, :DEFAULT, %i[custom_block])
+    @doc = QiitaMarker.render_doc(text, %i[UNSAFE], %i[custom_block])
     @expected = <<~HTML
       <div data-type="customblock" data-metadata="foo bar">
       <p>message</p>
+      <ul>
+      <li>list1</li>
+      <li>list2</li>
+      </ul>
+      <pre><code class="language-ruby">puts 'hello'
+      </code></pre>
+      <div>html block</div>
       </div>
     HTML
   end
 
   def test_to_html
-    assert_equal(@expected, @doc.to_html(:DEFAULT, %i[custom_block]))
+    assert_equal(@expected, @doc.to_html(%i[UNSAFE], %i[custom_block]))
   end
 end